feat: 검색 함수 분리
